feat : 시간 Long으로 수정

See merge request s09-final/S09P31B303!234

+++ greenstep/src/main/java/com/mm/greenstep/domain/plogging/service/MainService.java
@@ -21,7 +21,7 @@ public class MainService {
         List<Plogging> ploggingList = ploggingRepository.findAll();
 
         Double travelRange = 0.0;
-        Double travelTime  = 0.0;
+        Long travelTime  = 0L;
         Integer trashAmount  = 0;
 
         for (Plogging p : ploggingList) {
